Only archive weekly forecasts
We only want to archive forecasts weekly, not every time a change is made to the
codebase. This avoids lots of similar/identical forecasts resulting from active
development. Previously we were limiting archiving to once per day. This change
identifies builds on Travis that are part of the weekly runs and only archives
those.