Correctly state that prediction intervals are 90% in evaluation
The prediction intervals have always been 90%, but they were described and shown
as 95%. This fixes this issue and in the process creates a new metadata value
for the prediction interval (thanks to @ha0ye's suggestion) that can be used
throughout the code base. Updates to the models will be necessary to use this
value for all instances of prediction interval in the code base.