Make deploying results robust to PAT changes
GitHub recently updated it's PAT format which revealed a bug where
the PAT did not work properly because the remote was already present
and therefore adding it did not update the information. This fixes
the issue by removing the remote before adding it.

The error was also happening silently due to old uses of > /dev/null
to avoid leaking credentials on CI. Since we now running only on
HiPerGator these have been removed to make diagnosing future git errors
easier.